# lib 2/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Productivity Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List all habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    patch:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /habits/{id}/check:
    post:
      operationId: checkHabit
      tags: [habits]
      summary: Mark habit as done for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated habit with streak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"

  # ── FINANCES ───────────────────────────────────────────────────────────────
  /finances:
    get:
      operationId: listFinances
      tags: [finances]
      summary: List financial entries
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, all]
        - name: month
          in: query
          schema:
            type: string
            description: Filter by month (YYYY-MM)
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of financial entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceEntry"
    post:
      operationId: createFinanceEntry
      tags: [finances]
      summary: Create a financial entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceEntry"
  /finances/{id}:
    patch:
      operationId: updateFinanceEntry
      tags: [finances]
      summary: Update a financial entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceEntry"
    delete:
      operationId: deleteFinanceEntry
      tags: [finances]
      summary: Delete a financial entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /finances/monthly-summary:
    get:
      operationId: getMonthlySummary
      tags: [finances]
      summary: Get monthly spending breakdown by category
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            description: Month in YYYY-MM format
      responses:
        "200":
          description: Monthly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlySummary"

  # ── TAGS ───────────────────────────────────────────────────────────────────
  /tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all tags
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a custom tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── STATS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Habit:
      type: object
      required: [id, name, streak, currentStreak, level, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        streak:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        level:
          type: integer
        completedToday:
          type: boolean
        lastChecked:
          type: ["string", "null"]
        tagId:
          type: ["integer", "null"]
        createdAt:
          type: string
    HabitInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        tagId:
          type: integer
    HabitUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        tagId:
          type: integer

    # ── Finance schemas ───────────────────────────────────────────────────────
    FinanceEntry:
      type: object
      required: [id, type, amount, category, date, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        category:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        dueDate:
          type: ["string", "null"]
        isPaid:
          type: boolean
        tagId:
          type: ["integer", "null"]
        createdAt:
          type: string
    FinanceInput:
      type: object
      required: [type, amount, category, date]
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
        dueDate:
          type: string
        isPaid:
          type: boolean
        tagId:
          type: integer
    FinanceUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
        description:
          type: string
        date:
          type: string
        dueDate:
          type: string
        isPaid:
          type: boolean
        tagId:
          type: integer
    CategoryBreakdown:
      type: object
      required: [category, total, count]
      properties:
        category:
          type: string
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
    MonthlySummary:
      type: object
      required: [month, totalIncome, totalExpenses, netBalance, expensesByCategory]
      properties:
        month:
          type: string
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netBalance:
          type: number
        expensesByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        incomeByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"

    # ── Tag schemas ───────────────────────────────────────────────────────────
    Tag:
      type: object
      required: [id, name, color, isCustom, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        icon:
          type: ["string", "null"]
        isCustom:
          type: boolean
        createdAt:
          type: string
    TagInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        icon:
          type: string

    # ── Stats schemas ─────────────────────────────────────────────────────────
...